pyproject.toml
[project]
name = "opendesk"
version = "1.0.0"
description = "Open-source MCP server for AI desktop control. Works with Claude, Cursor, Ollama on Mac/Windows/Linux."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Mohan Kumar Swamynathan"}]
keywords = [
    "mcp", "model-context-protocol", "claude", "cursor", "ollama",
    "desktop-automation", "cross-platform", "local-ai", "ai-agent",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "mcp>=1.2.0",
    "psutil>=5.9.0",
    "pynput>=1.7.0",
    "Pillow>=10.0.0",
    "pyperclip>=1.8.0",
    "watchdog>=3.0.0",
    "rich>=13.0.0",
    "click>=8.1.0",
    "tomli>=2.0.0; python_version < '3.11'",
]

[project.optional-dependencies]
mac = ["pyobjc-framework-Cocoa>=9.0", "pyobjc-framework-Quartz>=9.0"]
windows = ["pywin32>=306", "pywinauto>=0.6.8", "pycaw>=20230407"]
linux = ["python-xlib>=0.33"]
ocr = ["pytesseract>=0.3.10"]
ollama = ["ollama>=0.1.0"]
dev = [
    "pytest>=7.0", "pytest-asyncio>=0.21", "pytest-cov>=4.0",
    "pytest-mock>=3.10", "pytest-timeout>=2.1",
    "ruff>=0.1.0", "mypy>=1.5",
]

[project.scripts]
opendesk = "opendesk.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["opendesk"]

[tool.pytest.ini_options]
timeout = 30
markers = [
    "mac: macOS-only tests",
    "windows: Windows-only tests",
    "linux: Linux-only tests",
    "slow: tests taking >5 seconds",
    "requires_display: tests needing display server",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = true
